Move desktop screen capture into modules/screen-capture and align preload build paths and imports.
